// src/client/services/service-proxies.ts
export interface WeightRange {
    _id: string;
    min: number;
    max: number;
    createdAt: number;
}
// ...
export interface Shop {
    _id: string;
    isActive: boolean;
    expiryDate: number;
    purchasedPackages: number[];
    createdAt: number;
    owner: any;
    domain: string;
    name: string;
    address: string;
    geocode: Geocode2;
    description: string;
    isFollowing: boolean;
    phone: string;
    email: string;
    representative: Representative2;
    introImages: string[];
    logoImage: string;
}

export interface ElasticSearchResult {
    data: ElasticSearchItemResult[];
    before: string;
    after: string;
}

export interface ElasticSearchOffsetResult {
    data: ElasticSearchItemResult[];
    total: number;
}

export interface ElasticSearchItemResult {
    title: string;
    state: string;
    description: string;
    shop: any;
    imageUrls: string[];
    price: number;
    priceType: string;
    newsType: string;
    owner: any;
    location: Location;
    status: string;
    usedHours: number;
    serialNo: number;
    categoryId: CategoryId;
    brand: Brand2;
    model: Model2;
    weight: number;
    year: number;
    followBy: string[];
    isDeleted: boolean;
    createdBy: number;
    createdAt: number;
    lastModifiedAt: number;
    elasticsearchID: string;
    elasticsearchTitle: string;
    isFollowing: boolean;
    _id: string;
}

export interface News {
    _id: string;
    title: string;
    state: string;
    description: string;
    imageUrls: string[];
    price: number;
    priceType: string;
    newsType: string;
    owner: any;
    shop: any;
    location: any;
    status: string;
    usedHours: number;
    serialNo: string;
    categoryId: any;
    type: any;
    brand: any;
    model: any;
    weight: number;
    year: number;
    weightRange: any;
    rentalPeriod: any;
    followBy: any[];
    isFollowing: boolean;
    createdAt: number;
    counterFollow: number;
    counterView: number;
}
// ...
export interface Brand {
    _id: string;
    slug: string;
    name: string;
    logoImage: string;
    totalNews: number;
    brandCategories: Category[];
    brandModels: Model[];
    createdAt: number;
}

export interface FindBrandsResult {
    data: Brand[];
    before: string;
    after: string;
}

export interface CreateBrandPayload {
    name: string;
    logoImage: string;
}

export interface UpdateBrandPayload {
    operation: UpdateBrandPayloadOperation;
    payload: any;
}

export interface Category {
    _id: string;
    slug: string;
    name: string;
    imageUrl: string;
    totalNews: number;
    categoryBrands: Brand[];
    categoryModels: Model[];
    createdAt: number;
}

export interface FindCategoriesResult {
    data: Category[];
}

export interface CreateCategoryPayload {
    name: string;
    imageUrl: string;
}

export interface UpdateCategoryPayload {
    operation: UpdateCategoryPayloadOperation;
    payload: Payload13;
}

export interface Model {
    _id: string;
    slug: string;
    name: string;
    brand: any;
    category: any;
    totalNews: number;
    createdAt: number;
}
// ...
export interface Geocode2 {
    lng: number;
    lat: number;
}

export interface Representative2 {
    fullName: string;
    identityNumber: string;
    email: string;
    paymentPhoneNo: string;
    address: string;
}

export interface Location {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
}

export interface CategoryId {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
}

export interface Brand2 {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
    totalNews: number;
}

export interface Model2 {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
    brand: string;
    category: string;
}